Start to create an RpcClient wrapper
